+++ src/Utils.h
@@ -404,7 +404,7 @@ namespace Ising_OPV {
 			diff[i] = fabs(data[i] - median);
 		}
 		auto it = min_element(diff.begin(), diff.end());
-		return distance(diff.begin(), it);
+		return (int)distance(diff.begin(), it);
 	}
 }
